backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Configuration
pydantic==2.7.1
pydantic-settings==2.2.1

# Document parsing
python-docx==1.1.2
pdfplumber==0.11.0

# LLM providers
openai==1.30.1
anthropic==0.26.0
httpx==0.27.0

# Encryption (for storing PM credentials)
cryptography==42.0.7

# PM integrations
jira==3.8.0
# asana, trello, azure-devops via httpx (REST)

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0
factory-boy==3.3.0
